refactor(misc1): move out autocmd related functions
