+++ src/nvim/ex_docmd.c
@@ -68,6 +68,7 @@
 #include "nvim/spell.h"
 #include "nvim/spellfile.h"
 #include "nvim/strings.h"
+#include "nvim/state.h"
 #include "nvim/syntax.h"
 #include "nvim/tag.h"
 #include "nvim/terminal.h"
